requirements.txt
selenium==4.20.0
pytest==8.2.0
pytest-html==4.1.1
webdriver-manager==4.0.2
allure-pytest==2.13.5
